LadderGame클래스로 generateResultsForAllPlayers 옮김
