+++ src/main/java/nextstep/ladder/domain/LadderGame.java
@@ -13,7 +13,25 @@ public class LadderGame {
 
     public LadderGame(Users users, Results results, List<LadderLine> lines) {
         gameInfo = new GameInfo(users, results);
-        this.ladder = new Ladder(lines);
+        ladder = new Ladder(lines);
+    }
+
+    public void generateResultsForAllPlayers() {
+        List<LadderLine> ladder = this.ladder.getLadder();
+        List<Result> results = gameInfo.getResults().getResults();
+
+        int index = 0;
+        for (User user : gameInfo.getUsers().getUsers()) {
+            user.setResult(results.get(getIndex(ladder, index)).getResult());
+            ++index;
+        }
+    }
+
+    private int getIndex(List<LadderLine> ladder, int index) {
+        for (LadderLine ladderLine : ladder) {
+            index = ladderLine.move(index);
+        }
+        return index;
     }
 
     public GameInfo getGameInfo() {
